Refuse a container with no identifier, and resolve its store once
`cloudKit(container:)` took whatever identifier the container had and fell
back to the literal "cloudKit" when it had none. That name is what the
store table is keyed by and what `Backend.id` remembers as the active
backend, so an absent identifier quietly became a made-up one. It throws
instead, which is a break in the signature — callers already inside a `try`
expression, the documented `setup(backends:)` snippets among them, need no
edit.

The store is resolved through `LazyTable`, which the factory now goes
through: one entry per identifier, built on first use, and everyone asking
for it waits on the same build. `NativeDatabase` holds the resolution
rather than the store, so nothing reaches the server until a read or a
write asks for it, and a backend stays as cheap to build as it looks.

The generic table has no idea what it holds. It keeps a `Task` per key and
hands out its value, so a value cannot be seen before whatever prepared it
finished — for CloudKit that preparation is publishing the entity catalog,
which is what a read used to await through a registration task of its own.

The rest is arrangement. `CatalogEntry` carries the entries it declares and
lives in its own file beside the extension holding them; `EntityCatalog`
keeps only what reads the catalog. The read helpers that took a store as a
parameter are extensions on it now.
