+++ Sources/Connectors/Native/NativeDatabase.swift
@@ -10,14 +10,7 @@ import Scout
 import ScoutDB
 
 struct NativeDatabase: Sendable {
-    let store: EntityStore
-
-    let registration: Task<Void, Never>
-
-    init(store: EntityStore, registration: Task<Void, Never> = Task {}) {
-        self.store = store
-        self.registration = registration
-    }
+    let resolve: @Sendable () async -> EntityStore
 }
 
 extension NativeDatabase: DatabaseWriter {
@@ -26,7 +19,7 @@ extension NativeDatabase: DatabaseWriter {
     }
 
     func write(records: [Record]) async throws {
-        await registration.value
+        let store = await resolve()
         for (entity, group) in Dictionary(grouping: records, by: \.recordType) {
             let batch = group.map { EntityWrite(values: Self.values(for: $0), uuid: $0.recordID) }
             try await store.write(batch, entity: entity)
@@ -42,14 +35,13 @@ extension NativeDatabase: DatabaseWriter {
 
 extension NativeDatabase: DatabaseReader {
     func read(matching query: RecordQuery, fields: [String]?) async throws -> RecordChunk {
-        await registration.value
+        let store = await resolve()
         let entity = query.recordType.recordType
         let sort = query.sort.first
 
-        // A keyset page orders by one field, and the store has no unbounded read
-        // to fall back on, so a query without a sort still walks in the order of
-        // the entity's own timestamp.
-        let records = try await builder(entity: entity, filters: query.filters).records(
+        let builder = query.filters.reduce(store.query(entity)) { $0.filter($1) }
+
+        let records = try await builder.records(
             orderedBy: sort?.field ?? EntityCatalog.dateField(for: entity),
             ascending: sort?.ascending ?? false
         )
@@ -58,16 +50,11 @@ extension NativeDatabase: DatabaseReader {
     }
 
     func read(matching query: RecordQuery, fields: [String]?, limit: Int) async throws -> RecordChunk {
-        await registration.value
-        let entity = query.recordType.recordType
-
-        // The keyset pager orders by a single field, so a query without a sort
-        // has no page order to follow — fall back to the full read.
         guard let sort = query.sort.first else {
             return try await read(matching: query, fields: fields)
         }
-        return try await page(
-            entity: entity,
+        return try await resolve().page(
+            entity: query.recordType.recordType,
             filters: query.filters,
             field: sort.field,
             ascending: sort.ascending,
@@ -75,40 +62,11 @@ extension NativeDatabase: DatabaseReader {
             after: nil
         )
     }
-
-    // Reads one keyset page and wraps its continuation cursor so the next page
-    // re-queries the store for the following bounded slice instead of retaining
-    // the whole pre-fetched tail in memory.
-    private func page(
-        entity: String, filters: [RecordQuery.Filter], field: String, ascending: Bool, limit: Int,
-        after cursor: FieldCursor?
-    ) async throws
-        -> RecordChunk
-    {
-        let page = try await builder(entity: entity, filters: filters)
-            .sort(field, ascending ? .forward : .reverse)
-            .page(size: limit, after: cursor)
-
-        return RecordChunk(
-            records: page.records.map(Record.init(entityRecord:)),
-            cursor: page.cursor.map { next in
-                RecordCursor { _ in
-                    try await self.page(
-                        entity: entity, filters: filters, field: field, ascending: ascending, limit: limit, after: next)
-                }
-            }
-        )
-    }
-
-    private func builder(entity: String, filters: [RecordQuery.Filter]) -> QueryBuilder {
-        filters.reduce(store.query(entity)) { $0.filter($1) }
-    }
 }
 
 extension NativeDatabase {
     func lookup(recordName: String, fields: [String]?) async throws -> Record {
-        await registration.value
-        guard let entityRecord = try await store.fetch(uuid: recordName) else {
+        guard let entityRecord = try await resolve().fetch(uuid: recordName) else {
             throw RecordNotFoundError()
         }
         return Record(entityRecord: entityRecord)
@@ -117,59 +75,47 @@ extension NativeDatabase {
 
 extension NativeDatabase {
     func series(matching query: SeriesQuery) async throws -> [MetricSeries] {
-        await registration.value
-        return try await NativeSeries(query: query).series(store: store)
+        try await NativeSeries(query: query).series(store: await resolve())
     }
 }
 
 extension NativeDatabase {
     func activity(in range: Range<Date>) async throws -> [ActivityPoint] {
-        await registration.value
-        // WAU/MAU windows reach back before the visible range.
+        let store = await resolve()
         let lookback = range.lowerBound.addingTimeInterval(-30 * .day).startOfDay
-
-        // Visit markers carry one record per device per day; sessions still
-        // back the days written before markers existed. The union dedupes, so
-        // dropping the session leg once every client ships markers is safe.
         let window = lookback..<range.upperBound
-        async let markers = visits(entity: VisitEntry.recordType, dateField: "date", in: window)
-        async let sessions = visits(entity: SessionEntry.recordType, dateField: "start_date", in: window)
+
+        async let markers = store.visits(
+            entity: VisitEntry.recordType,
+            dateField: "date",
+            in: window
+        )
+        async let sessions = store.visits(
+            entity: SessionEntry.recordType,
+            dateField: "start_date",
+            in: window
+        )
 
         return ActivityPoint.points(visits: try await markers + sessions, in: range)
-    }
-
-    private func visits(entity: String, dateField: String, in window: Range<Date>) async throws -> [ActivityVisit] {
-        try await datedIDs(entity: entity, dateField: dateField, idField: "device_id", in: window)
-            .map { ActivityVisit(date: $0.date, user: $0.id) }
-    }
-}
-
-extension NativeDatabase {
-    func datedIDs(entity: String, dateField: String, idField: String, in range: Range<Date>) async throws
-        -> [(date: Date, id: String)]
-    {
-        let records = try await store.records(entity: entity, dateField: dateField, in: range)
-
-        return records.compactMap { record -> (date: Date, id: String)? in
-            guard case .date(let date)? = record.values[dateField] else {
-                return nil
-            }
-            guard case .string(let id)? = record.values[idField] else {
-                return nil
-            }
-            return (date, id)
-        }
     }
 }
 
 extension NativeDatabase {
     func retention(in range: Range<Date>) async throws -> [RetentionCohort] {
-        await registration.value
+        let store = await resolve()
 
-        async let installs = datedIDs(
-            entity: InstallEntry.recordType, dateField: "date", idField: "install_id", in: range)
-        async let sessions = datedIDs(
-            entity: SessionEntry.recordType, dateField: "start_date", idField: "install_id", in: range)
+        async let installs = store.datedIDs(
+            entity: InstallEntry.recordType,
+            dateField: "date",
+            idField: "install_id",
+            in: range
+        )
+        async let sessions = store.datedIDs(
+            entity: SessionEntry.recordType,
+            dateField: "start_date",
+            idField: "install_id",
+            in: range
+        )
 
         var installDays: [String: Date] = [:]
         for install in try await installs {
@@ -181,6 +127,11 @@ extension NativeDatabase {
             sessionDays[session.id, default: []].insert(session.date.startOfDay)
         }
 
-        return RetentionCohort.build(installDays: installDays, sessionDays: sessionDays, in: range, asOf: Date())
+        return RetentionCohort.build(
+            installDays: installDays,
+            sessionDays: sessionDays,
+            in: range,
+            asOf: Date()
+        )
     }
 }
