+++ Sources/Connectors/Native/NativeReads.swift
@@ -6,14 +6,12 @@
 // https://opensource.org/licenses/MIT.
 
 import Foundation
+import Scout
 import ScoutDB
 
 let nativePageSize = 400
 
 extension QueryBuilder {
-    // The store bounds every read it offers, so a sweep is a walk over its
-    // keyset pages: one request per page rather than one for the whole entity,
-    // stopping as soon as a page runs short.
     func records(orderedBy field: String, ascending: Bool) async throws -> [EntityRecord] {
         let ordered = sort(field, ascending ? .forward : .reverse)
 
@@ -31,9 +29,66 @@ extension QueryBuilder {
 }
 
 extension EntityStore {
+    func page(
+        entity: String, filters: [RecordQuery.Filter], field: String, ascending: Bool, limit: Int,
+        after cursor: FieldCursor?
+    ) async throws -> RecordChunk {
+        let page = try await filters.reduce(query(entity)) { $0.filter($1) }
+            .sort(field, ascending ? .forward : .reverse)
+            .page(size: limit, after: cursor)
+
+        return RecordChunk(
+            records: page.records.map(Record.init(entityRecord:)),
+            cursor: page.cursor.map { next in
+                RecordCursor { _ in
+                    try await self.page(
+                        entity: entity,
+                        filters: filters,
+                        field: field,
+                        ascending: ascending,
+                        limit: limit,
+                        after: next
+                    )
+                }
+            }
+        )
+    }
+
     func records(entity: String, dateField: String, in range: Range<Date>) async throws -> [EntityRecord] {
         try await query(entity)
             .filter(dateField, in: range)
             .records(orderedBy: dateField, ascending: true)
+    }
+
+    func visits(entity: String, dateField: String, in window: Range<Date>) async throws -> [ActivityVisit] {
+        try await datedIDs(
+            entity: entity,
+            dateField: dateField,
+            idField: "device_id",
+            in: window
+        )
+        .map {
+            ActivityVisit(date: $0.date, user: $0.id)
+        }
+    }
+
+    func datedIDs(entity: String, dateField: String, idField: String, in range: Range<Date>) async throws
+        -> [(date: Date, id: String)]
+    {
+        let records = try await records(
+            entity: entity,
+            dateField: dateField,
+            in: range
+        )
+
+        return records.compactMap { record -> (date: Date, id: String)? in
+            guard case .date(let date)? = record.values[dateField] else {
+                return nil
+            }
+            guard case .string(let id)? = record.values[idField] else {
+                return nil
+            }
+            return (date, id)
+        }
     }
 }
